job_seeker_service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.7
pydantic==1.10.7
psycopg2-binary==2.9.5
requests==2.28.2
python-dotenv==1.0.0
pytest==7.3.1
httpx==0.24.0